+++ tests/test_checker/test_noqa.py
@@ -108,6 +108,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS235': 1,
     'WPS236': 1,
     'WPS237': 1,
+    'WPS238': 1,
 
     'WPS300': 1,
     'WPS301': 1,