+++ wemake_python_styleguide/violations/complexity.py
@@ -58,7 +58,7 @@ Summary
    TooManyImportedModuleMembersViolation
    TooLongTupleUnpackViolation
    TooComplexFormattedStringViolation
-
+   TooManyRaisesViolation
 
 Module complexity
 -----------------
@@ -100,6 +100,7 @@ Structure complexity
 .. autoclass:: TooManyImportedModuleMembersViolation
 .. autoclass:: TooLongTupleUnpackViolation
 .. autoclass:: TooComplexFormattedStringViolation
+.. autoclass:: TooManyRaisesViolation
 
 """
 
@@ -1139,11 +1140,12 @@ class TooManyImportedModuleMembersViolation(ASTViolation):
         Default:
         :str:`wemake_python_styleguide.options.defaults.MAX_IMPORT_FROM_MEMBERS`
 
-    .. versionadded:: 0.14.0
+    .. versionadded:: 0.15.0
 
     """
 
     error_template = 'Found too many imported names from a module: {0}'
+
     code = 235
 
 
@@ -1219,3 +1221,31 @@ class TooComplexFormattedStringViolation(ASTViolation):
 
     error_template = 'Found a too complex `f` string'
     code = 237
+
+
+@final
+class TooManyRaisesViolation(ASTViolation):
+    """
+    Forbids too many ``raise`` statements in a function.
+
+    Reasoning:
+        Too many ``raise`` statements in a function make the code
+        untraceable and overcomplicated.
+
+    Solution:
+        Split the function into smaller functions, such that
+        each of them can raise less errors.
+        Create more standard errors, or use alternative ways to
+        raise them.
+
+    Configuration:
+        This rule is configurable with ``--max-raises``.
+        Default:
+        :str:`wemake_python_styleguide.options.defaults.MAX_RAISES`
+
+    .. versionadded:: 0.15.0
+
+    """
+
+    error_template = 'Found too many raises in a function: {0}'
+    code = 238